refactor: Provide padding from patterns to components (#228)
Refactor `LeftAlignedScreen` and components so that components are no
longer responsible for providing the correct horizontal padding.

After this change, `LeftAlignedScreen` supplies the horizontal padding
to components through slot parameters for components to apply.

This is a pattern used by
[`Scaffold`](https://developer.android.com/develop/ui/compose/components/scaffold).

It allows our components to have a built-in padding that is aligned with
stock components such as `Text` and `Image` (that is, zero padding).
This in turn provides consistency in the way components behave and
allows our components to be used in the same way as stock components at
a given use site.

DCMAW-8054