fix: Update typography to match Figma (#248)
- Update typography to match how Figma displays text
- Ensure that the `LocalTextStyle` provided by `GdsTheme` is identical
to _Body Large_

Initially we discovered that the default text style provided by
`GdsTheme` didn't match our _Body Large_ style. As an example, the
following two `Text`s would not be identical as they would be under
`MaterialTheme`:

```kt
GdsTheme {
    Text(
        text = "Using the theme default style"
    )
    Text(
        text = "Explicitly setting the default text style",
        style = MaterialTheme.typography.bodyLarge
    )
}
````

This is because when the `MaterialTheme` creates the `LocalTextStyle`,
it merges our `Typography` definition with a default text style that
includes a 'line height style' and a 'platform style'.

To fix the inconsistency between `LocalTextStyle.current` and
`MaterialTheme.typography.bodyLarge`, we would need to configure a
`lineHeightStyle` and `platformStyle` for all our text styles.

Further research led me to discover that these settings were added to so
that developers could easily configure their typography to align to the
way Figma presents text.

* https://medium.com/androiddevelopers/fixing-font-padding-in-compose-text-768cd232425b
* https://dev.to/canyudev/android-and-figma-typography-and-how-to-achieve-100-fidelity-l40

The settings presented in the articles, that help align typography with
Figma designs, are what I've used in this PR.