feat: Add custom back handler for `FullScreenDialogue` (#254)
[Tutorial for writing good descriptions]: https://cbea.ms/git-commit/

[//]: # 

#[ Add custom back handler for
`FullScreenDialogue`](https://govukverify.atlassian.net/browse/DCMAW-11964)

- When using the `FullScreenDialogue` in the `LocalAuthImplementation` -
we require custom back behaviour. Based on the FullScreenDialogue
implementation it looks like the dismiss is always required which would
technically have the same behaviour as pressing the device back button.
We have talked about this in MP and overall the back button behaviour
would always align with the dismiss - however, please do raise any
potential issues if there are cases/ situations I might have missed/ not
account for - **This should not be a breaking change as it's not
changing the API, rather than assigning the `onDismissRequest` to the
`BackHandler`**

[//]: # (e.g. "- Create 'androidLibrary' Gradle module.")

## Evidence of the change
No evidence to show - check code

## Checklist

### Before creating the pull request

- [x] Commit messages that conform to conventional commit messages.
- [x] Tests pass locally.
- [x] Pull request has a clear title with a short description about the
feature or update.

### Before the CODEOWNERS review the pull request

- [x] Self-review code.
- [x] Successfully run changes on a testing device.
- [x] Complete automated Testing:
  * [x] Unit Tests.
  * [x] Integration Tests.
  * [x] Instrumentation / Emulator Tests.
- [ ] Handle PR comments.

### Before merging the pull request

- [ ] [Sonar cloud report] passes inspections for your PR.
- [ ] Resolve all comments.

[Sonar cloud report]:
https://sonarcloud.io/project/overview?id=di-mobile-android-ui
[Accessibility considerations]:
https://developer.android.com/guide/topics/ui/accessibility/testing